feat: implementa PedidoCompraService

+++ src/main/java/com/gestaooserp/dev/repository/PedidoCompraRepository.java
@@ -3,5 +3,5 @@ package com.gestaooserp.dev.repository;
 import com.gestaooserp.dev.entity.PedidoCompra;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PedidoCompraRepository extends JpaRepository<PedidoCompra,Integer> {
+public interface PedidoCompraRepository extends JpaRepository<PedidoCompra,Long> {
 }
